requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.8.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
boto3==1.17.69
botocore==1.20.112
certifi==2021.10.8
charset-normalizer==2.0.9
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.4.0
django-jquery==3.1.0
django-localflavor==3.1
django-nested-admin==4.1.1
django-npm==1.0.0
django-phonenumber-field==7.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-smart-selects==1.5.9
django-storages==1.11.1
docutils==0.17.1
et-xmlfile==1.1.0
Faker==19.3.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.26.4
openpyxl==3.1.2
packaging==21.3
pandas==2.2.2
phonenumbers==8.13.39
Pillow==9.2.0
platformdirs==2.2.0
psycopg2-binary==2.9.1
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.4
python-monkey-business==1.1.0
python-stdnum==1.16
pytz==2021.1
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
snowballstemmer==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.6
virtualenv==20.4.7
xlrd==2.0.1
